requirements.txt
alembic==1.14.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mysqlclient==2.2.6
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-magic==0.4.27
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
